pyproject.toml
[project]
name = "apex-trading"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "alpaca-py>=0.43.2",
    "anthropic>=0.84.0",
    "apscheduler>=3.11.2",
    "duckdb>=1.4.4",
    "httpx>=0.28.1",
    "lightgbm>=4.6.0",
    "numpy>=2.4.2",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pyarrow>=23.0.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "streamlit>=1.55.0",
    "vectorbt>=0.28.4",
    "yfinance>=1.2.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
